// packages/nger-core/lib/sdk/sdk_types.ts
export interface AccountInfo {
    /** 小程序帐号信息 */
    miniProgram: MiniProgram;
    /** 插件帐号信息（仅在插件中调用时包含这一项） */
    plugin: Plugin;
}
// ...
export interface BackgroundAudioManager {
    /** 音频已缓冲的时间，仅保证当前播放时间点到此时间点内容已缓冲。（只读） */
    buffered: number;
    /** 封面图 URL，用于做原生音频播放器背景图。原生音频播放器中的分享功能，分享出去的卡片配图及背景也将使用该图。 */
    coverImgUrl: string;
    /** 当前音频的播放位置（单位：s），只有在有合法 src 时返回。（只读） */
    currentTime: number;
    /** 当前音频的长度（单位：s），只有在有合法 src 时返回。（只读） */
    duration: number;
    /** 专辑名，原生音频播放器中的分享功能，分享出去的卡片简介，也将使用该值。 */
    epname: string;
    /** 当前是否暂停或停止。（只读） */
    paused: boolean;
    /** 音频协议。默认值为 'http'，设置 'hls' 可以支持播放 HLS 协议的直播音频。
     *
     * 最低基础库： `1.9.94` */
    protocol: string;
    /** 歌手名，原生音频播放器中的分享功能，分享出去的卡片简介，也将使用该值。 */
    singer: string;
    /** 音频的数据源（{% version('2.2.3') %} 开始支持云文件ID）。默认为空字符串，**当设置了新的 src 时，会自动开始播放**，目前支持的格式有 m4a, aac, mp3, wav。 */
    src: string;
    /** 音频开始播放的位置（单位：s）。 */
    startTime: number;
    /** 音频标题，用于原生音频播放器音频标题（必填）。原生音频播放器中的分享功能，分享出去的卡片标题，也将使用该值。 */
    title: string;
    /** 页面链接，原生音频播放器中的分享功能，分享出去的卡片简介，也将使用该值。 */
    webUrl: string;
}
// ...
export interface CanvasContext {
    /** 填充颜色。用法同 [CanvasContext.setFillStyle()]。
     *
     * 最低基础库： `1.9.90` */
    fillStyle: string;
    /** 当前字体样式的属性。符合 [CSS font 语法](https://developer.mozilla.org/zh-CN/docs/Web/CSS/font) 的 DOMString 字符串，至少需要提供字体大小和字体族名。默认值为 10px sans-serif。
     *
     * 最低基础库： `1.9.90` */
    font: string;
    /** 全局画笔透明度。范围 0-1，0 表示完全透明，1 表示完全不透明。 */
    globalAlpha: number;
    /** 在绘制新形状时应用的合成操作的类型。目前安卓版本只适用于 `fill` 填充块的合成，用于 `stroke` 线段的合成效果都是 `source-over`。
     *
     * 目前支持的操作有
     * - 安卓：xor, source-over, source-atop, destination-out, lighter, overlay, darken, lighten, hard-light
     * - iOS：xor, source-over, source-atop, destination-over, destination-out, lighter, multiply, overlay, darken, lighten, color-dodge, color-burn, hard-light, soft-light, difference, exclusion, saturation, luminosity
     *
     * 最低基础库： `1.9.90` */
    globalCompositeOperation: string;
    /** 线条的端点样式。用法同 [CanvasContext.setLineCap()]。
     *
     * 最低基础库： `1.9.90` */
    lineCap: number;
    /** 虚线偏移量，初始值为0
     *
     * 最低基础库： `1.9.90` */
    lineDashOffset: number;
    /** 线条的交点样式。用法同 [CanvasContext.setLineJoin()]。
     *
     * 最低基础库： `1.9.90` */
    lineJoin: number;
    /** 线条的宽度。用法同 [CanvasContext.setLineWidth()]。
     *
     * 最低基础库： `1.9.90` */
    lineWidth: number;
    /** 最大斜接长度。用法同 [CanvasContext.setMiterLimit()]。
     *
     * 最低基础库： `1.9.90` */
    miterLimit: number;
    /** 阴影的模糊级别
     *
     * 最低基础库： `1.9.90` */
    shadowBlur: number;
    /** 阴影的颜色
     *
     * 最低基础库： `1.9.90` */
    shadowColor: number;
    /** 阴影相对于形状在水平方向的偏移
     *
     * 最低基础库： `1.9.90` */
    shadowOffsetX: number;
    /** 阴影相对于形状在竖直方向的偏移
     *
     * 最低基础库： `1.9.90` */
    shadowOffsetY: number;
    /** 边框颜色。用法同 [CanvasContext.setFillStyle()]。
     *
     * 最低基础库： `1.9.90` */
    strokeStyle: string;
}
// ...
export interface CreateAnimationOption {
    /** 动画延迟时间，单位 ms */
    delay?: number;
    /** 动画持续时间，单位 ms */
    duration?: number;
    /** 动画的效果
     *
     * 可选值：
     * - 'linear': 动画从头到尾的速度是相同的;
     * - 'ease': 动画以低速开始，然后加快，在结束前变慢;
     * - 'ease-in': 动画以低速开始;
     * - 'ease-in-out': 动画以低速开始和结束;
     * - 'ease-out': 动画以低速结束;
     * - 'step-start': 动画第一帧就跳至结束状态直到结束;
     * - 'step-end': 动画一直保持开始状态，最后一帧跳到结束状态; */
    timingFunction?:
    | 'linear'
    | 'ease'
    | 'ease-in'
    | 'ease-in-out'
    | 'ease-out'
    | 'step-start'
    | 'step-end';
    transformOrigin?: string;
}
export interface CreateBLEConnectionOption {
    /** 用于区分设备的 id */
    deviceId: string;
    /** 接口调用结束的回调函数（调用成功、失败都会执行） */
    complete?: CreateBLEConnectionCompleteCallback;
    /** 接口调用失败的回调函数 */
    fail?: CreateBLEConnectionFailCallback;
    /** 接口调用成功的回调函数 */
    success?: CreateBLEConnectionSuccessCallback;
    /** 超时时间，单位ms，不填表示不会超时 */
    timeout?: number;
}
/** 选项 */
export interface CreateIntersectionObserverOption {
    /** 初始的相交比例，如果调用时检测到的相交比例与这个值不相等且达到阈值，则会触发一次监听器的回调函数。 */
    initialRatio?: number;
    /** 是否同时观测多个目标节点（而非一个），如果设为 true ，observe 的 targetSelector 将选中多个节点（注意：同时选中过多节点将影响渲染性能）
     *
     * 最低基础库： `2.0.0` */
    observeAll?: boolean;
    /** 一个数值数组，包含所有阈值。 */
    thresholds?: Array<number>;
}
// ...
export interface DownloadFileOption {
    /** 下载资源的 url */
    url: string;
    /** 接口调用结束的回调函数（调用成功、失败都会执行） */
    complete?: DownloadFileCompleteCallback;
    /** 接口调用失败的回调函数 */
    fail?: DownloadFileFailCallback;
    /** 指定文件下载后存储的路径
     *
     * 最低基础库： `1.8.0` */
    filePath?: string;
    /** HTTP 请求的 Header，Header 中不能设置 Referer */
    header?: object;
    /** 接口调用成功的回调函数 */
    success?: DownloadFileSuccessCallback;
}
// ...
export interface ExtInfo {
    /** 第三方平台自定义的数据 */
    extConfig: object;
}
// ...
export interface GetBatteryInfoSyncResult {
    /** 是否正在充电中 */
    isCharging: boolean;
    /** 设备电量，范围 1 - 100 */
    level: string;
}
// ...
export interface GetLogManagerOption {
    /** 取值为0/1，取值为0表示是否会把 `App`、`Page` 的生命周期函数和 `wx` 命名空间下的函数调用写入日志，取值为1则不会。默认值是 0
     *
     * 最低基础库： `2.3.2` */
    level?: number;
}
// ...
export interface GetStorageInfoSyncOption {
    /** 当前占用的空间大小, 单位 KB */
    currentSize: number;
    /** 当前 storage 中所有的 key */
    keys: Array<string>;
    /** 限制的空间大小，单位 KB */
    limitSize: number;
}
// ...
export interface GetSystemInfoSyncResult {
    /** 客户端基础库版本
     *
     * 最低基础库： `1.1.0` */
    SDKVersion: string;
    /** 允许微信使用相册的开关（仅 iOS 有效）
     *
     * 最低基础库： `2.6.0` */
    albumAuthorized: boolean;
    /** 设备性能等级（仅Android小游戏）。取值为：-2 或 0（该设备无法运行小游戏），-1（性能未知），>=1（设备性能值，该值越高，设备性能越好，目前最高不到50）
     *
     * 最低基础库： `1.8.0` */
    benchmarkLevel: number;
    /** 蓝牙的系统开关
     *
     * 最低基础库： `2.6.0` */
    bluetoothEnabled: boolean;
    /** 设备品牌
     *
     * 最低基础库： `1.5.0` */
    brand: string;
    /** 允许微信使用摄像头的开关
     *
     * 最低基础库： `2.6.0` */
    cameraAuthorized: boolean;
    /** 用户字体大小（单位px）。以微信客户端「我-设置-通用-字体大小」中的设置为准
     *
     * 最低基础库： `1.5.0` */
    fontSizeSetting: number;
    /** 微信设置的语言 */
    language: string;
    /** 允许微信使用定位的开关
     *
     * 最低基础库： `2.6.0` */
    locationAuthorized: boolean;
    /** 地理位置的系统开关
     *
     * 最低基础库： `2.6.0` */
    locationEnabled: boolean;
    /** 允许微信使用麦克风的开关
     *
     * 最低基础库： `2.6.0` */
    microphoneAuthorized: boolean;
    /** 设备型号 */
    model: string;
    /** 允许微信通知带有提醒的开关（仅 iOS 有效）
     *
     * 最低基础库： `2.6.0` */
    notificationAlertAuthorized: boolean;
    /** 允许微信通知的开关
     *
     * 最低基础库： `2.6.0` */
    notificationAuthorized: boolean;
    /** 允许微信通知带有标记的开关（仅 iOS 有效）
     *
     * 最低基础库： `2.6.0` */
    notificationBadgeAuthorized: boolean;
    /** 允许微信通知带有声音的开关（仅 iOS 有效）
     *
     * 最低基础库： `2.6.0` */
    notificationSoundAuthorized: boolean;
    /** 设备像素比 */
    pixelRatio: number;
    /** 客户端平台 */
    platform: string;
    /** 屏幕高度，单位px
     *
     * 最低基础库： `1.1.0` */
    screenHeight: number;
    /** 屏幕宽度，单位px
     *
     * 最低基础库： `1.1.0` */
    screenWidth: number;
    /** 状态栏的高度，单位px
     *
     * 最低基础库： `1.9.0` */
    statusBarHeight: number;
    /** 操作系统及版本 */
    system: string;
    /** 微信版本号 */
    version: string;
    /** Wi-Fi 的系统开关
     *
     * 最低基础库： `2.6.0` */
    wifiEnabled: boolean;
    /** 可使用窗口高度，单位px */
    windowHeight: number;
    /** 可使用窗口宽度，单位px */
    windowWidth: number;
}
// ...
export interface InnerAudioContext {
    /** 是否自动开始播放，默认为 `false` */
    autoplay: boolean;
    /** 音频缓冲的时间点，仅保证当前播放时间点到此时间点内容已缓冲（只读） */
    buffered: number;
    /** 当前音频的播放位置（单位 s）。只有在当前有合法的 src 时返回，时间保留小数点后 6 位（只读） */
    currentTime: number;
    /** 当前音频的长度（单位 s）。只有在当前有合法的 src 时返回（只读） */
    duration: number;
    /** 是否循环播放，默认为 `false` */
    loop: boolean;
    /** 是否遵循系统静音开关，默认为 `true`。当此参数为 `false` 时，即使用户打开了静音开关，也能继续发出声音。从 2.3.0 版本开始此参数不生效，使用 `wx.setInnerAudioOption` 接口统一设置。 */
    obeyMuteSwitch: boolean;
    /** 当前是是否暂停或停止状态（只读） */
    paused: boolean;
    /** 音频资源的地址，用于直接播放。{% version('2.2.3') %} 开始支持云文件ID */
    src: string;
    /** 开始播放的位置（单位：s），默认为 0 */
    startTime: number;
    /** 音量。范围 0~1。默认为 1
     *
     * 最低基础库： `1.9.90` */
    volume: number;
}
// ...
export interface LaunchOptionsApp {
    /** 启动小程序的路径 */
    path: string;
    /** 启动小程序的 query 参数 */
    query: object;
    /** 来源信息。从另一个小程序、公众号或 App 进入小程序时返回。否则返回 `{}`。(参见后文注意) */
    referrerInfo: LaunchOptionsAppReferrerInfo;
    /** 启动小程序的[场景值] */
    scene: number;
    /** shareTicket，详见[获取更多转发信息] */
    shareTicket: string;
}
// ...
export interface MiniProgram {
    /** 小程序 appId */
    appId: string;
}
// ...
export interface Rect {
    /** 下边界坐标，单位：px */
    bottom: number;
    /** 高度，单位：px */
    height: number;
    /** 左边界坐标，单位：px */
    left: number;
    /** 右边界坐标，单位：px */
    right: number;
    /** 上边界坐标，单位：px */
    top: number;
    /** 宽度，单位：px */
    width: number;
}
// ...
export interface AudioContext {
    /** [AudioContext.pause()](AudioContext.pause.md)
     *
     * 暂停音频。 */
    pause(): void;
    /** [AudioContext.play()](AudioContext.play.md)
     *
     * 播放音频。 */
    play(): void;
    /** [AudioContext.seek(number position)](AudioContext.seek.md)
     *
     * 跳转到指定位置。 */
    seek(
        /** 跳转位置，单位 s */
        position: number,
    ): void;
    /** [AudioContext.setSrc(string src)](AudioContext.setSrc.md)
     *
     * 设置音频地址 */
    setSrc(
        /** 音频地址 */
        src: string,
    ): void;
}
// ...
export interface CameraContext {
    /** [CameraContext.startRecord(Object object)](CameraContext.startRecord.md)
     *
     * 开始录像 */
    startRecord(option: CameraContextStartRecordOption): void;
    /** [CameraContext.stopRecord(Object object)](CameraContext.stopRecord.md)
     *
     * 结束录像 */
    stopRecord(option?: StopRecordOption): void;
    /** [CameraContext.takePhoto(Object object)](CameraContext.takePhoto.md)
     *
     * 拍摄照片 */
    takePhoto(option: TakePhotoOption): void;
}
// ...
export interface DownloadTask {
    /** [DownloadTask.abort()](DownloadTask.abort.md)
     *
     * 中断下载任务
     *
     * 最低基础库： `1.4.0` */
    abort(): void;
    /** [DownloadTask.offHeadersReceived(function callback)](DownloadTask.offHeadersReceived.md)
     *
     * 取消监听 HTTP Response Header 事件
     *
     * 最低基础库： `2.1.0` */
    offHeadersReceived(
        /** HTTP Response Header 事件的回调函数 */
        callback: DownloadTaskOffHeadersReceivedCallback,
    ): void;
    /** [DownloadTask.offProgressUpdate(function callback)](DownloadTask.offProgressUpdate.md)
     *
     * 取消监听下载进度变化事件
     *
     * 最低基础库： `2.1.0` */
    offProgressUpdate(
        /** 下载进度变化事件的回调函数 */
        callback: DownloadTaskOffProgressUpdateCallback,
    ): void;
    /** [DownloadTask.onHeadersReceived(function callback)](DownloadTask.onHeadersReceived.md)
     *
     * 监听 HTTP Response Header 事件。会比请求完成事件更早
     *
     * 最低基础库： `2.1.0` */
    onHeadersReceived(
        /** HTTP Response Header 事件的回调函数 */
        callback: DownloadTaskOnHeadersReceivedCallback,
    ): void;
    /** [DownloadTask.onProgressUpdate(function callback)](DownloadTask.onProgressUpdate.md)
     *
     * 监听下载进度变化事件
     *
     * 最低基础库： `1.4.0` */
    onProgressUpdate(
        /** 下载进度变化事件的回调函数 */
        callback: DownloadTaskOnProgressUpdateCallback,
    ): void;
}
// ...
export interface LivePlayerContext {
    /** [LivePlayerContext.exitFullScreen(Object object)](LivePlayerContext.exitFullScreen.md)
     *
     * 退出全屏 */
    exitFullScreen(option?: ExitFullScreenOption): void;
    /** [LivePlayerContext.mute(Object object)](LivePlayerContext.mute.md)
     *
     * 静音 */
    mute(option?: MuteOption): void;
    /** [LivePlayerContext.pause(Object object)](LivePlayerContext.pause.md)
     *
     * 暂停
     *
     * 最低基础库： `1.9.90` */
    pause(option?: LivePlayerContextPauseOption): void;
    /** [LivePlayerContext.play(Object object)](LivePlayerContext.play.md)
     *
     * 播放 */
    play(option?: PlayOption): void;
    /** [LivePlayerContext.requestFullScreen(Object object)](LivePlayerContext.requestFullScreen.md)
     *
     * 进入全屏 */
    requestFullScreen(option: LivePlayerContextRequestFullScreenOption): void;
    /** [LivePlayerContext.resume(Object object)](LivePlayerContext.resume.md)
     *
     * 恢复
     *
     * 最低基础库： `1.9.90` */
    resume(option?: LivePlayerContextResumeOption): void;
    /** [LivePlayerContext.stop(Object object)](LivePlayerContext.stop.md)
     *
     * 停止 */
    stop(option?: LivePlayerContextStopOption): void;
}
// ...
const query = wx.createSelectorQuery()